A glider submenu module.
