+++ main/setup/SetupMenu.cpp
@@ -8,6 +8,7 @@
 #include "setup/SetupMenu.h"
 #include "setup/SubMenuDevices.h"
 #include "setup/SubMenuCompassWind.h"
+#include "setup/SubMenuGlider.h"
 #include "setup/ShowBootMsg.h"
 #include "screen/SetupRoot.h"
 #include "IpsDisplay.h"
@@ -41,9 +42,7 @@
 #include <inttypes.h>
 #include <iterator>
 #include <algorithm>
-// #include <cstring>
 #include <string>
-#include "SetupMenu.h"
 
 static void setup_create_root(SetupMenu *top);
 
@@ -61,9 +60,6 @@ static void audio_menu_create_tonestyles(SetupMenu *top);
 static void audio_menu_create_deadbands(SetupMenu *top);
 static void audio_menu_create_equalizer(SetupMenu *top);
 static void audio_menu_create_mute(SetupMenu *top);
-
-static void glider_menu_create(SetupMenu *top);
-static void glider_menu_create_polarpoints(SetupMenu *top);
 
 static void options_menu_create(SetupMenu *top);
 static void options_menu_create_units(SetupMenu *top);
@@ -334,23 +330,6 @@ int factv_adj(SetupMenuValFloat *p) {
 	float bat = getBattery()->get(true);
 	MYUCG->setPrintPos(1, 100);
 	MYUCG->printf("%0.2f Volt", bat);
-	return 0;
-}
-
-int polar_select(SetupMenuSelect *p) {
-	ESP_LOGI(FNAME,"glider-index %d", p->getValue());
-	glider_type_index.set(p->getValue());
-	p->getParent()->setBuzzword(p->value());
-	return 0;
-}
-
-static int start_weight_adj(SetupMenuValFloat *p) {
-	float wingload = (ballast_kg.get() + crew_weight.get() + empty_weight.get()) / polar_wingarea.get();
-	ESP_LOGI(FNAME,"wingload:%.1f empty: %.1f cw:%.1f water:%.1f", wingload, empty_weight.get(), crew_weight.get(), ballast_kg.get() );
-	MYUCG->setFont(ucg_font_fub25_hr, true);
-	MYUCG->setPrintPos(1, 110);
-	MYUCG->printf("%0.2f kg/m2  ", wingload);
-	MYUCG->setFont(ucg_font_ncenR14_hr);
 	return 0;
 }
 
@@ -958,61 +937,6 @@ void audio_menu_create(SetupMenu *audio) {
 	afac->setHelp(
 			"How the audio frequency responds to the climb rate: < 1 for logarithmic, and > 1 for exponential, response");
 	audio->addEntry(afac);
-}
-
-void glider_menu_create_polarpoints(SetupMenu *top) {
-	SetupMenuValFloat *wil = new SetupMenuValFloat("Ref Wingload", "kg/m2", nullptr, false, &polar_wingload);
-	wil->setHelp(
-			"Wingloading that corresponds to the 3 value pairs for speed/sink of polar");
-	top->addEntry(wil);
-	SetupMenuValFloat *pov1 = new SetupMenuValFloat("Speed 1", "km/h", nullptr, false, &polar_speed1);
-	pov1->setHelp("Speed 1, near minimum sink from polar e.g. 80 km/h");
-	top->addEntry(pov1);
-	SetupMenuValFloat *pos1 = new SetupMenuValFloat("Sink  1", "m/s", nullptr, false, &polar_sink1);
-	top->addEntry(pos1);
-	SetupMenuValFloat *pov2 = new SetupMenuValFloat("Speed 2", "km/h", nullptr, false, &polar_speed2);
-	pov2->setHelp("Speed 2 for a moderate cruise from polar e.g. 120 km/h");
-	top->addEntry(pov2);
-	SetupMenuValFloat *pos2 = new SetupMenuValFloat("Sink  2", "m/s", nullptr, false, &polar_sink2);
-	top->addEntry(pos2);
-	SetupMenuValFloat *pov3 = new SetupMenuValFloat("Speed 3", "km/h", nullptr, false, &polar_speed3);
-	pov3->setHelp("Speed 3 for a fast cruise from polar e.g. 170 km/h");
-	top->addEntry(pov3);
-	SetupMenuValFloat *pos3 = new SetupMenuValFloat("Sink  3", "m/s", nullptr, false, &polar_sink3);
-	top->addEntry(pos3);
-}
-
-void glider_menu_create(SetupMenu *poe) {
-	SetupMenuSelect *glt = new SetupMenuSelect("Type", RST_NONE, polar_select, &glider_type_index);
-	poe->addEntry(glt);
-	ESP_LOGI(FNAME, "#polars %d", Polars::numPolars());
-	for (int x = 0; x < Polars::numPolars(); x++) {
-		ESP_LOGI(FNAME, "P: %s - %d", Polars::getPolarName(x), Polars::getPolarIndex(x));
-		glt->addEntry(Polars::getPolarName(x), Polars::getPolarIndex(x));
-	}
-	poe->setHelp("Weight and polar setup for best match with performance of glider");
-	glt->setSelect(Polars::getGliderEnumPos()); // Cannot set select from nvs variable value containing the magix index
-	ESP_LOGI(FNAME, "Number of Polars installed: %d", Polars::numPolars() );
-
-	SetupMenu *pa = new SetupMenu("Polar Points", glider_menu_create_polarpoints);
-	pa->setHelp("Adjust the polar at 3 points, in the commonly used metric system",230);
-	poe->addEntry(pa);
-
-	SetupMenuValFloat *maxbal = new SetupMenuValFloat("Max Ballast", "liters", nullptr, false, &polar_max_ballast);
-	poe->addEntry(maxbal);
-
-	SetupMenuValFloat *wingarea = new SetupMenuValFloat("Wing Area", "m2", nullptr, false, &polar_wingarea);
-	poe->addEntry(wingarea);
-
-	SetupMenuValFloat *fixball = new SetupMenuValFloat("Empty Weight", "kg", start_weight_adj, false, &empty_weight);
-	fixball->setPrecision(0);
-	fixball->setHelp("Net rigged weight of the glider, according to the weight and balance plan");
-	fixball->setNeverInline();
-	poe->addEntry(fixball);
-
-	SetupMenuValFloat *vmax = new SetupMenuValFloat("Maximum Speed", "", nullptr, false, &v_max);
-	vmax->setHelp("Configure maximum speed for corresponding aircraft type");
-	poe->addEntry(vmax);
 }
 
 void options_menu_create_units(SetupMenu *top) {
