+++ main/setup/SubMenuGlider.h
@@ -1,0 +1,8 @@
+
+#pragma once
+
+class SetupMenu;
+class SetupMenuValFloat;
+
+void glider_menu_create(SetupMenu *top);
+int start_weight_adj(SetupMenuValFloat *p);
